Site 17/06

+++ src/main/java/br/itb/projeto/evobody/model/entity/Usuario.java
@@ -1,3 +1,6 @@
+
+
+
 package br.itb.projeto.evobody.model.entity;
 
 import javax.persistence.Entity;
@@ -12,7 +15,7 @@ public class Usuario {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	private String nome;
 	private String email;
 	private String senha;
@@ -23,7 +26,7 @@ public class Usuario {
 	public long getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getNome() {
